refactor: reduce code in totalNutrition, better compose files
