+++ app/dashboard/@calendar/page.tsx
@@ -1,5 +1,5 @@
 import CalendarView from "@/components/Dashboard/CalendarView";
-import TotalNutrition from "@/components/TotalNutrition/TotalNutrition";
+import TotalNutrition from "@/components/Dashboard/TotalNutrition";
 
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
