+++ components/Dashboard/MealsTable/MealsTable.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table";
-import MealsTableBody from '../MealsTableBody/MealsTableBody';
 import { getServerSession } from 'next-auth';
+import MealsTableBody from './MealsTableBody/MealsTableBody';
 
 const MealsTable = async () => {
   const session = await getServerSession();
